Refactor BigInteger and BigDecimal handling in SQLDA
Implemented changes in BigInteger and BigDecimal handling in SQLDA within the fbclient library. "Int128" type removed and instead native 'LongArray' is used to represent 128-bit integers. BigInteger and BigDecimal processing methods are now better documented and simplified for clarity. This should improve the ease of use when interacting with large numeric types.